Refactoring ObjectDetection main detect method
